:recycle: refactor: koreainvestment
